refactor(staking): rearchitect flows to make components more reusable
