--- src/containers/staking/StakeFlow/steps/Stake/Stake.module.scss
@@ -1,43 +1,0 @@
-.Container {
-	> * {
-		&:not(:first-child) {
-			margin-top: 24px;
-		}
-	}
-
-	.Back {
-		padding: 0;
-		background: none;
-		color: var(--color-text-primary);
-
-		img {
-			margin-right: 8px;
-			height: 24px;
-			width: 24px;
-		}
-	}
-
-	.Message {
-		display: flex;
-		align-items: center;
-		width: 100%;
-		padding: 16px;
-
-		background: var(--color-purple-lighter);
-
-		.Content {
-			flex: 1;
-			margin-left: 16px;
-		}
-	}
-
-	.Stats {
-		> * {
-			flex: 1;
-
-			&:not(:first-of-type) {
-				margin-left: 16px;
-			}
-		}
-	}
-}
