+++ src/containers/staking/flows/StakeFlow/StakeFlow.tsx
@@ -27,7 +27,6 @@ const StakeFlow = (props: StakeFlowProps) => {
 	const [step, setStep] = useState<Steps>(Steps.Stake);
 
 	const onStake = (amount: number) => {
-		console.log(amount);
 		setStep(Steps.Approve);
 	};
 
@@ -44,7 +43,6 @@ const StakeFlow = (props: StakeFlowProps) => {
 						poolDomain={stakingOn.domain}
 						onBack={onClose}
 						onStake={onStake}
-						isTransactionPending
 					/>
 				);
 			case Steps.Approve:
