+++ src/containers/staking/flows/StakeFlow/steps/Approve/Approve.module.scss
