+++ src/containers/staking/flows/components/PoolData/PoolData.module.scss
@@ -1,0 +1,9 @@
+.Stats {
+	> * {
+		flex: 1;
+
+		&:not(:first-of-type) {
+			margin-left: 16px;
+		}
+	}
+}
