+++ src/containers/staking/flows/components/index.ts
@@ -1,0 +1,2 @@
+export { default as PoolData } from './PoolData/PoolData';
+export { default as Message } from './Message/Message';
