+++ src/containers/staking/index.ts
@@ -11,4 +11,5 @@ export { default as StakePool } from './StakePool/StakePool';
 
 export { default as StakeModule } from './StakeModule/StakeModule';
 
-export { default as StakeFlow } from './StakeFlow/StakeFlow';
+export { default as StakeFlow } from './flows/StakeFlow/StakeFlow';
+export { default as ClaimFlow } from './flows/ClaimFlow/ClaimFlow';
